+++ tests/plugins/xcube/test_plugin.py
@@ -14,6 +14,9 @@ class ExportPluginTest(TestCase):
                 "grid-mapping-naming",
                 "increasing-time",
                 "lat-lon-naming",
+                "ml-dataset-meta",
+                "ml-dataset-time",
+                "ml-dataset-xy",
                 "single-grid-mapping",
                 "time-naming",
             },
@@ -32,9 +35,9 @@ class ExportPluginTest(TestCase):
         all_rule_names = set(f"xcube/{k}" for k in plugin.rules.keys())
         self.assertEqual(
             all_rule_names,
-            set(plugin.configs["all"].rules.keys()),
+            set(plugin.configs["all"][-1].rules.keys()),
         )
         self.assertEqual(
             all_rule_names,
-            set(plugin.configs["recommended"].rules.keys()),
+            set(plugin.configs["recommended"][-1].rules.keys()),
         )