+++ xrlint/plugins/xcube/rules/lat_lon_naming.py
@@ -1,6 +1,6 @@
 from xrlint.node import DatasetNode
 from xrlint.plugins.xcube.constants import LAT_NAME, LON_NAME
-from xrlint.plugins.xcube.rules import plugin
+from xrlint.plugins.xcube.plugin import plugin
 from xrlint.rule import RuleContext, RuleOp
 
 INVALID_LAT_NAMES = {"ltd", "latitude"}